rename all files to follow camelCase and not camel.case
