+++ src/actions/showTypes.js
