+++ src/components/Filter.js
@@ -12,7 +12,7 @@ import 'date-fns';
 import Grid from '@material-ui/core/Grid';
 import DateFnsUtils from '@date-io/date-fns';
 import { MuiPickersUtilsProvider, KeyboardDatePicker } from '@material-ui/pickers';
-import useStyles from './Filter.styles';
+import useStyles from './FilterStyles';
 import CurrentFilter from './CurrentFilter';
 
 // prettier-ignore
