+++ src/components/Loader.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import CircularProgress from '@material-ui/core/CircularProgress';
-import useStyles from './Loader.styles';
+import useStyles from './LoaderStyles';
 
 const Loader = () => {
   const classes = useStyles();
