+++ src/components/NavBar.js
@@ -5,7 +5,7 @@ import Toolbar from '@material-ui/core/Toolbar';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 
-import useStyles from './NavBar.styles';
+import useStyles from './NavBarStyles';
 
 // prettier-ignore
 const NavBar = () => {
