+++ src/components/ShowCardStyles.js
