+++ src/pages/AllShowsPage.js
@@ -2,9 +2,9 @@ import React from 'react';
 import PropTypes from 'prop-types';
 import Pagination from '@material-ui/lab/Pagination';
 import Button from '@material-ui/core/Button';
-import useStyles from './AllShowsPage.styles';
+import useStyles from './AllShowsPageStyles';
 import ShowCard from '../components/ShowCard';
-import Filter from '../components/Filter.container';
+import Filter from '../components/FilterContainer';
 
 // prettier-ignore
 const AllShowPage = ({
